refactor: typescript folder restructure
