+++ src/ads/banner/index.ts
@@ -1,0 +1,2 @@
+export { default as BannerAd } from './BannerAd';
+export { default as GAMBannerAd } from './GAMBannerAd';
