+++ src/ads/fullscreen/index.ts
@@ -1,0 +1,4 @@
+export { default as InterstitialAd } from './InterstitialAd';
+export { default as RewardedAd } from './RewardedAd';
+export { default as RewardedInterstitialAd } from './RewardedInterstitialAd';
+export { default as AppOpenAd } from './AppOpenAd';
